Add upload tool.
Currently the tool is just a skeleton to test the functionality. It will
evolve gradually by adding functionality. For now it only uploads the
file in the appropriate place. It uses the configuration file to know
the server/path pair from it. In the future it will be able to update
and add entries in the configuration file.