Remove some of the TODOs
Handling errors will require implementing logger and as this is
development tool we can leave for now the exceptions which are
informative enough to explain what the problem is.

In the future we can make some generic gem which exports the logger
functionality out of launcher and share it with all our other gems.