Try to guess relative to config file path.
If the file appended is in a subdirectory relative to the config file
then pass only the relative part to be stored in the configuration,
otherwise fails. The latter will be handled in the future by may be
allowing a switch to specify the exact key. Till then such entries can
be added by hand.